refactor: move models to src/ package
