+++ api/src/main/kotlin/io/github/slaxnetwork/api/dto/game/cookieclicker/CookieClickerProfile.kt
@@ -1,4 +1,4 @@
-package io.github.slaxnetwork.api.models.views.profile.game.cookieclicker
+package io.github.slaxnetwork.api.dto.game.cookieclicker
 
 import kotlinx.serialization.Serializable
