+++ src/main/kotlin/io/github/slaxnetwork/database/impl/postgres/repositories/PostgresProfileRepository.kt
@@ -1,9 +1,9 @@
 package io.github.slaxnetwork.database.impl.postgres.repositories
 
 import com.github.jasync.sql.db.SuspendingConnection
-import io.github.slaxnetwork.api.models.rank.RankModel
-import io.github.slaxnetwork.api.models.profile.game.GameProfileModel
-import io.github.slaxnetwork.api.models.profile.ProfileModel
+import io.github.slaxnetwork.database.models.rank.RankModel
+import io.github.slaxnetwork.database.models.profile.game.GameProfileModel
+import io.github.slaxnetwork.database.models.profile.ProfileModel
 import io.github.slaxnetwork.database.impl.postgres.utils.execute
 import io.github.slaxnetwork.database.impl.postgres.utils.firstNullableRow
 import io.github.slaxnetwork.database.impl.postgres.utils.firstRow
