+++ src/main/kotlin/io/github/slaxnetwork/database/impl/postgres/repositories/PostgresRanksRepository.kt
@@ -1,7 +1,7 @@
 package io.github.slaxnetwork.database.impl.postgres.repositories
 
 import com.github.jasync.sql.db.SuspendingConnection
-import io.github.slaxnetwork.api.models.rank.RankModel
+import io.github.slaxnetwork.database.models.rank.RankModel
 import io.github.slaxnetwork.database.impl.postgres.utils.execute
 import io.github.slaxnetwork.database.impl.postgres.utils.firstNullableRow
 import io.github.slaxnetwork.database.impl.postgres.utils.firstRow
