+++ src/main/kotlin/io/github/slaxnetwork/database/models/profile/game/cookieclicker/CookieClickerUpgradesModel.kt
@@ -1,8 +1,8 @@
-package io.github.slaxnetwork.api.models.profile.game.cookieclicker
+package io.github.slaxnetwork.database.models.profile.game.cookieclicker
 
 import com.github.jasync.sql.db.RowData
 import io.github.slaxnetwork.api.exceptions.DatabaseDeserializeException
-import io.github.slaxnetwork.api.models.views.profile.game.cookieclicker.CookieClickerUpgrades
+import io.github.slaxnetwork.api.dto.game.cookieclicker.CookieClickerUpgrades
 
 data class CookieClickerUpgradesModel(
     val testUpgrade: Int
